requirements.txt
pymongo
tornado
motor
numpy
scipy
pillow
